DW-545: Make element work in section, page etc.
