--- web/modules/custom/os2forms_organisation/os2forms_organisation.module
@@ -1,18 +1,0 @@
-<?php
-
-/**
- * @file
- * Module file for the os2forms_organisation module.
- */
-
-use Drupal\Core\Form\FormStateInterface;
-use Drupal\os2forms_organisation\Helper\FormHelper;
-
-/**
- * Implements hook_form_alter().
- *
- * Allows altering of forms.
- */
-function os2forms_organisation_form_alter(&$form, FormStateInterface $form_state, $form_id) {
-  Drupal::service(FormHelper::class)->formAlter($form, $form_state, $form_id);
-}
